fix gcc

+++ cpp/models/ode_secirvvs/parameters_io.cpp
@@ -224,7 +224,8 @@ IOResult<void> read_confirmed_cases_data(
     return success();
 }
 
-IOResult<std::vector<std::vector<double>>> read_immunity_population(const std::string& path, const size_t& num_age_groups)
+IOResult<std::vector<std::vector<double>>> read_immunity_population(const std::string& path,
+                                                                    const size_t& num_age_groups)
 {
     std::vector<std::vector<double>> ans(3, std::vector<double>(num_age_groups, 0.0));
     std::fstream immunity_file;
@@ -241,7 +242,7 @@ IOResult<std::vector<std::vector<double>>> read_immunity_population(const std::s
             if (linenumber < 2)
                 tp.erase(tp.size() - 1);
             auto line = split(tp, ' ');
-            for (int i = 0; i < num_age_groups; i++) {
+            for (size_t i = 0; i < num_age_groups; i++) {
                 ans[linenumber][i] = std::stod(line[i]);
             }
             linenumber++;
@@ -383,17 +384,20 @@ IOResult<void> set_vaccination_data(std::vector<Model>& model, const std::string
         auto date_df = vacc_data_entry.date;
         if (it != vregion.end()) {
             auto region_idx = size_t(it - vregion.begin());
-            AgeGroup age        = vacc_data_entry.age_group;
+            AgeGroup age    = vacc_data_entry.age_group;
 
             // initialize the temporary immunity states
             if (date_df >=
                     offset_date_by_days(
-                        date, static_cast<int>(
-                        -model[region_idx].parameters.template get<TimeTemporaryImmunityPI>()[age] -
+                        date,
+                        static_cast<int>(
+                            -model[region_idx].parameters.template get<TimeTemporaryImmunityPI>()[age] -
                             model[region_idx].parameters.template get<DaysUntilEffectiveImprovedImmunity>()[age])) &&
                 date_df <=
                     offset_date_by_days(
-                        date, static_cast<int>(-model[region_idx].parameters.template get<DaysUntilEffectiveImprovedImmunity>()[age]))) {
+                        date,
+                        static_cast<int>(
+                            -model[region_idx].parameters.template get<DaysUntilEffectiveImprovedImmunity>()[age]))) {
                 model[region_idx].populations[{age, InfectionState::TemporaryImmunPartialImmunity}] +=
                     vacc_data_entry.num_vaccinations_completed;
             }
@@ -401,11 +405,13 @@ IOResult<void> set_vaccination_data(std::vector<Model>& model, const std::string
             if (date_df >=
                     offset_date_by_days(
                         date,
-                        static_cast<int>(-model[region_idx].parameters.template get<TimeTemporaryImmunityII>()[age] -
+                        static_cast<int>(
+                            -model[region_idx].parameters.template get<TimeTemporaryImmunityII>()[age] -
                             model[region_idx].parameters.template get<DaysUntilEffectiveBoosterImmunity>()[age])) &&
                 date_df <=
                     offset_date_by_days(
-                        date, static_cast<int>(-model[region_idx].parameters.template get<DaysUntilEffectiveBoosterImmunity>()[age]))
+                        date, static_cast<int>(
+                                  -model[region_idx].parameters.template get<DaysUntilEffectiveBoosterImmunity>()[age]))
 
             ) {
                 model[region_idx].populations[{age, InfectionState::TemporaryImmunImprovedImmunity}] +=
